refactor: 임시 로그인 버튼 header로 변경
